requirements.txt
docker==5.0.3
fastapi[all]==0.74.1
postgres==4.0
aerich==0.6.3
tortoise-orm[asyncpg]==0.19.0
passlib[bycrypt]==1.7.4
python-jose[cryptography]==3.3.0

redis==4.2.2